refactor: restructure controller files and update import paths for consistency
