+++ src/controllers/index.ts
@@ -1,4 +1,4 @@
-export * as projects from './Projects';
-export * as tags from './Tags';
-export * as tasks from './Tasks';
-export * as statuses from './Statuses';
+export * as projects from './project.controller';
+export * as tags from './tag.controller';
+export * as tasks from './task.controller';
+export * as statuses from './status.controller';
